+++ src/analyzer.ts
@@ -4,6 +4,9 @@ import {
   File,
   LocationInFile,
 } from './types';
+
+import { cleanRelativePath } from './parser/util';
+
 export { Analysis } from './types';
 
 interface FileExport {
@@ -65,9 +68,28 @@ const getExportMap = (files: File[]): ExportMap => {
   return map;
 };
 
+const processImportsOfExportedAsNamespace = (
+  file: File,
+  exportMap: ExportMap,
+): void => {
+  /* Basic support for export-as-namespace.
+   * If a file is exported as a namespace, and that namespace is imported,
+   * then we mark *all* exports of that file as used.
+   * A more accurate analysis would require scanning for all usages of the exports, via that namespace.
+   */
+  file.pathsExportedAsNamespace.forEach(exportedAsNamespace => {
+    const what = exportMap[exportedAsNamespace]?.exports;
+    if (what) {
+      Object.keys(what).forEach(exported => what[exported].usageCount++);
+    }
+  });
+};
+
 const processImports = (file: File, exportMap: ExportMap): void => {
+  processImportsOfExportedAsNamespace(file, exportMap);
+
   Object.keys(file.imports).forEach(key => {
-    let ex = exportMap[key] && exportMap[key].exports;
+    let ex = exportMap[key]?.exports;
 
     // Handle imports from an index file
     if (!ex && key === '.') {
@@ -118,7 +140,7 @@ const expandExportFromStar = (files: File[], exportMap: ExportMap): void => {
       .forEach(ex => {
         delete fileExports.exports[ex];
 
-        const exports = exportMap[ex.slice(2)]?.exports;
+        const exports = exportMap[cleanRelativePath(ex)]?.exports;
         if (exports) {
           Object.keys(exports)
             .filter(e => e != 'default')